Reduce provider runtime-loader transport coupling
Move provider HTTP transport/error classification and shared record reading out of runtime-loader.ts so the main runtime module stays focused on provider request shaping, response parsing, and runtime factory orchestration. Keep the public runtime-loader export surface stable by re-exporting the same typed provider errors.

Constraint: Behavior-preserving refactor only; no runtime API changes or new dependencies
Rejected: Split OpenAI/Anthropic/Google builders in the same pass | would widen the slice beyond one low-risk responsibility move
Rejected: Add new provider-http unit tests first | existing runtime-loader transient error tests already lock the moved behavior end-to-end
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep provider transport/error mapping isolated from request-body builders so future provider-specific rewrites can stay small and testable
Tested: VF_DISABLE_LRU_INTERVAL=1 deno test --no-check --allow-all src/provider/runtime-loader.test.ts src/provider/runtime-loader/provider-request-init.test.ts; deno fmt --check src/provider/runtime-loader.ts src/provider/runtime-loader/provider-http.ts src/provider/runtime-loader/provider-records.ts; DENO_NO_PACKAGE_JSON=1 deno lint src/provider/runtime-loader.ts src/provider/runtime-loader/provider-http.ts src/provider/runtime-loader/provider-records.ts; deno check src/provider/runtime-loader.ts src/provider/runtime-loader/provider-http.ts src/provider/runtime-loader/provider-records.ts; deno task typecheck
Not-tested: Full deno task test suite before push; pre-push hook is expected to run it
