+++ src/provider/runtime-loader/provider-records.ts
@@ -1,0 +1,7 @@
+export function readRecord(value: unknown): Record<string, unknown> | undefined {
+  if (!value || typeof value !== "object" || Array.isArray(value)) {
+    return undefined;
+  }
+
+  return Object.fromEntries(Object.entries(value));
+}
